move its into hadoop catalog it
